Switched all of the prime number libraries to using int64

+++ euler/003.go
@@ -5,7 +5,7 @@ import "math"
 const Problem003Default = 600851475143
 
 func Problem003(toFactor int64) int64 {
-	limit := int(math.Sqrt(float64(toFactor)) + 0.5)
+	limit := int64(math.Sqrt(float64(toFactor)) + 0.5)
 	primes := PrimesUnder(limit)
 	for i := len(primes) - 1; i > 0; i-- {
 		if toFactor%int64(primes[i]) == 0 {
